Import of ATRICKETT/Log-Trivial-0.20 from CPAN.
gitpan-cpan-distribution: Log-Trivial
gitpan-cpan-version:      0.20
gitpan-cpan-path:         ATRICKETT/Log-Trivial-0.20.tar.gz
gitpan-cpan-author:       ATRICKETT
gitpan-cpan-maturity:     released

+++ t/04-posix-write.t
@@ -2,11 +2,13 @@
 
 use strict;
 use Test;
-BEGIN { plan tests => 22 }
+use Fcntl qw(:DEFAULT);
+BEGIN { plan tests => 24 }
 
 use Log::Trivial;
 
-my $logfile="./t/test.log";
+my $logfile = "./t/test.log";
+my $o_sync  = "./t/o_sync";
 ok(1);
 
 #	2-6
@@ -18,30 +20,61 @@ ok($logger);
 ok($logger->set_log_file($logfile));				# Set the test file to read
 ok($logger->{_file}, $logfile);
 
-#	5-6
+#	7-8
 ok($logger->set_log_mode("m"));						# Set to multi/slow mode
 ok($logger->{_mode}, 1);
 
-#	7-9
+#	9-11
 ok($logger->{_level}, 3);							# Check the default level 3
 ok($logger->set_log_level(2));						# Set the logging level to 2
 ok($logger->{_level}, 2);							# Check it's set
 
-#	10-14
+#	12-17
 ok(! -e $logfile);										# There should be no file there now
 ok(! $logger->write(comment => "Test", level => 3));	# Write Test to the log, shouldn't be written
-ok($logger->write(comment => "Test-m", level => 1));	# Write Test to the log, should be written
-ok($logger->write("Test-2-m"));							# Write without a level
-ok(-e $logfile);										# Now there should be a file
 
-#	15-16
-ok($logger->set_log_mode("s"));							# Set to single/fast mode
-ok($logger->{_mode}, 0);
+eval {
+    sysopen my $log, $o_sync, O_WRONLY | O_CREAT | O_SYNC | O_APPEND;
+};
+if ($@ =~ /Your vendor has not defined Fcntl macro O_SYNC/) {
+print STDERR <<"WARNING";
 
-#	17-18
-ok(! $logger->write(comment => "Test", level => 3));	# Write Test to the log, shouldn't be written
-ok($logger->write(comment => "Test-s", level => 1));	# Write Test to the log, should be written
+#################################################
+# This module uses the POSIX open O_SYNC flag.  #
+# This flag is not supported on this system.    #
+# You may still use this module, but not in     #
+# O_SYNC mode, please see the docs for details. #
+#################################################
 
-#	19-20
-ok(unlink $logfile); 
+WARNING
+
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+skip("Non POSIX Platform");
+
+} else {
+    ok(unlink $o_sync);
+    ok($logger->write(comment => "Test-m", level => 1)); 	# Write Test to the log, should be written
+
+    ok($logger->write("Test-2-m"));							# Write without a level
+    ok(-e $logfile);										# Now there should be a file
+
+    #	17-18
+    ok($logger->set_log_mode("s"));							# Set to single/fast mode
+    ok($logger->{_mode}, 0);
+
+    #	19-20
+    ok(! $logger->write(comment => "Test", level => 3));	# Write Test to the log, shouldn't be written
+    ok($logger->write(comment => "Test-s", level => 1));	# Write Test to the log, should be written
+
+    #	21-22
+    ok(unlink $logfile);
+};
 ok(! -e $logfile);
+ok(! -e $o_sync);
